Landing new error explanations code. Now uses OpenSP's message IDs to catch each emitted message and a Config::General config file to store custom explanations. An intended future development is context-sensitive hints and tips, l10n/i18n, and overriding of the messages themselves.
For now it shows the original error message and merely includes the
verbose explanation. A future enhancement would be to supress the
original error message alltogether and display our own customized
version (which is possibly localized based on Accept-Language headers).

A minimal script for parsing the OpenSP message file is included and
should be relatively easy to modify to support migrating error
explanations and other value-add data if and when OpenSP updates.

The major unresolved points remaining revolve around presentation:

  * Should explanations be triggered by the Verbose switch?
    Or should they merely become more verbose with it enabled?

  * Can we use JavaScript to display explanations on demand?
    (Jim is working on it, but we'll see whether he can come
     up with something that will scale well (both up and down)).

  * How best to encourage users to submit improved explanations?

  * Is it appropriate to include the "Help Wanted" spiels in the
    production release?

Otherwise this code should be reasonably solid.